+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 2.6.1 (2025-10-01)
 - fix a few issues with the new touch placement feature
+- add a workaround for Zwift Click V2 which resets the device when button events are no longer sent
 
 ### 2.6.0 (2025-09-30)
 - refactor touch placements: show touches on screen, fix misplaced coordinates - should fix #64